Split a new `CommandState` out from `CommandDesc`.
This makes `CommandDesc` a stateless, reusable representation for a command to
be cached, and delays resolving any application state - such as the cwd or any
environment variables - until the command is ready to be executed.

Most callers shouldn't need to directly interact with `CommandState` as it's
still possible to pass a reference to a `CommandDesc` to `Bkt`. Unlike
`CommandDesc`, a `CommandState` instance is a single-shot object that is
consumed when passed to `Bkt`. Typically `CommandState` objects will be very
short-lived, whereas `CommandDesc` objects can be persisted and invoked
repeatedly.

This separation will enable keying off a file's modtime, as discussed in #10.