Ignore NotFound errors when deleting expired data.
When starting bkt in parallel, the multiple bkts race to clean up
expired data. This opens the possibility that deletions will fail
because the file has already been deleted... but that's okay, we wanted
to do that anyway. So this change ignores NotFound errors on the
deletion.

Currently, the result of this race is errors as follows:

```
bkt: Cache lookup failed: Failed to remove expired data: No such file or directory (os error 2)
```

Fixes #38