Implement streaming support for subprocess output.
As suggested in https://stackoverflow.com/q/66060139 the child process'
streams are processed in background threads responsible for both
persisting and streaming the output to the caller's stdout/stderr.

Scoped threads are used to support streaming to references such as
`&mut Vec<u8>` and to loosen the type requirements of the stream (namely
to not require that they are 'static).

CLI:
* Subprocess out/err streams are duplicated to bkt's out/err while the
  subprocess runs, instead of caching all output first and then writing
  the cached output after the subprocess completes.
* Streaming is enabled by default and is not configurable.

Library:
* Added `Bkt::retrieve_streaming` and `Bkt::refresh_streaming` methods
  which accept out/err sinks. The subprocess' output is written to these
  streams in addition to being cached and available in the returned
  `Invocation`.
* These methods are marked "experimental" and are subject to change or
  even removal.