Improve handling of truncated streams
DisregardBrokenPipe needs to implement write_all() as well as write()
because the former treats a 0-byte write as an error, whereas DPB should
no-op in that scenario.

Note that this is fix is in main.rs and therefore _only_ fixes the CLI.
Along with the existing TODO on DPB it may be preferable to stop using
write_all() in lib.rs, but for now I'm erring on the side of letting
library users handle failures themselves.

Fixes #44