Cache directories are now user-specific
Adjust the cache directory naming scheme to include the EUID, thereby
allowing multiple users on the same machine to invoke `bkt` without
encountering permission issues trying to access the same directory.

For now this only applies to Unix, which is also the only platform where
bkt attempts to restrict access to the cache directory.

Also refreshed the docs on BKT_TMPDIR / BKT_CACHE_DIR.

Added some best-effort test coverage of this behavior; it's hard to reliably
test with multiple users that the test has access to, so for now it
simply tries to use sudo and silently skips the test if that fails.

Fixes #35